+++ changelog.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 <!-- Add all new changes here. They will be moved under a version at release -->
-`2024-11-26`
+* `FIX` missing-fields diagnostic now warns about missing inherited fields
 * `CHG` Update Love2d version
 
 ## 3.13.2